LayoutTests/imported/w3c/web-platform-tests/css/css-overscroll-behavior/overscroll-behavior-logical-expected.txt


FAIL Logical overscroll-behavior maps correctly when element has horizontal-tb writing mode assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Logical overscroll-behavior maps correctly when element has vertical-rl writing mode assert_equals: expected (string) "contain" but got (undefined) undefined
FAIL Logical overscroll-behavior maps correctly when element has horizontal-tb writing mode and is not affected by rtl direction assert_equals: expected (string) "none" but got (undefined) undefined
